+++ ext/ManifoldsBaseRecursiveArrayToolsExt/ManifoldsBaseRecursiveArrayToolsExt.jl
@@ -9,7 +9,7 @@ if isdefined(Base, :get_extension)
 
     import ManifoldsBase:
         allocate,
-        allocate_as,
+        allocate_on,
         allocate_result,
         default_inverse_retraction_method,
         default_retraction_method,
@@ -41,7 +41,7 @@ else
 
     import ..ManifoldsBase:
         allocate,
-        allocate_as,
+        allocate_on,
         allocate_result,
         default_inverse_retraction_method,
         default_retraction_method,