+++ test/power.jl
@@ -90,22 +90,22 @@ end
         allocate(M, p) isa Vector{SMatrix{2,2,Float64,4}}
     end
 
-    @testset "allocate_as" begin
+    @testset "allocate_on" begin
         M = ManifoldsBase.DefaultManifold(2, 2)
         N = PowerManifold(M, NestedReplacingPowerRepresentation(), 2)
-        p = allocate_as(N)
+        p = allocate_on(N)
         @test p isa Vector{Matrix{Float64}}
         @test size(p) == (2,)
 
-        p = allocate_as(N, Vector{Matrix{Float32}})
+        p = allocate_on(N, Vector{Matrix{Float32}})
         @test p isa Vector{Matrix{Float32}}
         @test size(p) == (2,)
 
-        X = allocate_as(N, TangentSpaceType())
+        X = allocate_on(N, TangentSpaceType())
         @test X isa Vector{Matrix{Float64}}
         @test size(X) == (2,)
 
-        X = allocate_as(N, TangentSpaceType(), Vector{Matrix{Float32}})
+        X = allocate_on(N, TangentSpaceType(), Vector{Matrix{Float32}})
         @test X isa Vector{Matrix{Float32}}
         @test size(X) == (2,)
     end