+++ NEWS.md
@@ -5,11 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.15.11] 16/07/2024
+## [0.15.11] 23/07/2024
 
 ### Added
 
-* Function `allocate_as` to generically allocate point and tangent vectors on a manifold without a pre-existing instance but of a particular type.
+* Function `allocate_on` to generically allocate point and tangent vectors on a manifold without a pre-existing instance but of a particular type.
 * Function `default_type` to get the default type of points and tangent vectors for a manifold.
 
 ## [0.15.10] 19/05/2024